### 2017-04-16 #239 @whedon
[PRE REVIEW]: IBCAO_py: A matplotlib library for using the International Bathymetric Chart of the Arctic Ocean with cartopy and matplotlib
**Submitting author:** @gauteh (<a href="http://orcid.org/0000-0002-5653-1447">Gaute Hope</a>)
**Repository:** <a href="https://github.com/gauteh/ibcao_py" target ="_blank">https://github.com/gauteh/ibcao_py</a>
**Version:** v1.0
