### 2017-04-17 #239 @arfon
@Mankoff - would you be willing to review this submission for JOSS?
