### 2017-04-24 #239 @dvalters
@arfon if not I would be happy to - quite interested in cartopy/matplotib things.
