### 2017-04-24 #239 @arfon
> @arfon if not I would be happy to - quite interested in cartopy/matplotib things.

👍 great!
