### 2017-04-24 #239 @arfon
@dvalters please head over to #250 to carry out the actual review.
