### 2017-04-24 9249137
Add setup tools, modularize and move COPYING to LICENSE
