### 2017-05-03 #250 @gauteh
Arfon Smith writes on april 24, 2017 22:57:
> @dvalters - please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: http://joss.theoj.org/about#reviewer_guidelines
>

@dvalters: is your review done? I do not get any notifications on changes
to checkboxes.

Best regards, Gaute
