### 2017-05-03 gauteh/ibcao_py#1 @dvalters
Warnings about unclosed files when running test script
I received the following warnings about unclosed files when running the unit test script (`runtests.sh`):

(It may be nothing serious, but just thought you should be aware. It's not a requirement to fix this for acceptance in JOSS openjournals/joss-reviews#250)

```
INFO:root:testing depth against gmt extracted depths
setting up interpolation function..
ibcao: closing map.
/home/dav/anaconda3/envs/cartopy/lib/python3.6/site-packages/scipy/io/netcdf.py:299: RuntimeWarning: Cannot close a netcdf_file opened with mmap=True, when netcdf_variables or arrays referring to its data still exist. All data arrays obtained from such files refer directly to data on disk, and must be copied before the file can be cleanly closed. (See netcdf_file docstring for more information on mmap.)
  ), category=RuntimeWarning)
ok
test_depths (test_depth.IbcaoDepthTest) ... ibcao read, shape: (11617, 11617)
INFO:root:calculating depth profile
setting up interpolation function..
INFO:root:d=
INFO:root:(10000,)
ibcao: closing map.
ok
test_known_positions (test_depth.IbcaoDepthTest) ... ibcao read, shape: (11617, 11617)
INFO:root:testing depth on a few known positions
setting up interpolation function..
ibcao: closing map.
/home/dav/anaconda3/envs/cartopy/lib/python3.6/site-packages/scipy/io/netcdf.py:299: RuntimeWarning: Cannot close a netcdf_file opened with mmap=True, when netcdf_variables or arrays referring to its data still exist. All data arrays obtained from such files refer directly to data on disk, and must be copied before the file can be cleanly closed. (See netcdf_file docstring for more information on mmap.)
  ), category=RuntimeWarning)
ok
test_map_coordinates (test_depth_interp.IbcaoDepthTest) ... ibcao read, shape: (11617, 11617)
INFO:root:testing map coordinates
setting up interpolation function..
INFO:root:dz=
INFO:root:(10000,)
ibcao: closing map.
/home/dav/anaconda3/envs/cartopy/lib/python3.6/site-packages/scipy/io/netcdf.py:299: RuntimeWarning: Cannot close a netcdf_file opened with mmap=True, when netcdf_variables or arrays referring to its data still exist. All data arrays obtained from such files refer directly to data on disk, and must be copied before the file can be cleanly closed. (See netcdf_file docstring for more information on mmap.)
  ), category=RuntimeWarning)
ok
test_rect_bivariate_spline (test_depth_interp.IbcaoDepthTest) ... ibcao read, shape: (11617, 11617)
INFO:root:testing rectbivariatespline
INFO:root:running map_depth..
ibcao: closing map.
/home/dav/anaconda3/envs/cartopy/lib/python3.6/site-packages/scipy/io/netcdf.py:299: RuntimeWarning: Cannot close a netcdf_file opened with mmap=True, when netcdf_variables or arrays referring to its data still exist. All data arrays obtained from such files refer directly to data on disk, and must be copied before the file can be cleanly closed. (See netcdf_file docstring for more information on mmap.)
  ), category=RuntimeWarning)
ok
test_resample_depth (test_depth_map.IbcaoDepthTest) ... ibcao read, shape: (11617, 11617)
INFO:root:testing resampling of depth
INFO:root:resampling to: (59, 59)
INFO:root:interpolation done
/home/dav/anaconda3/envs/cartopy/lib/python3.6/site-packages/cartopy/feature.py:184: ResourceWarning: unclosed file <_io.BufferedReader name='/home/dav/.local/share/cartopy/shapefiles/natural_earth/physical/10m_coastline.shp'>
  geometries = tuple(shapereader.Reader(path).geometries())
/home/dav/anaconda3/envs/cartopy/lib/python3.6/site-packages/cartopy/feature.py:184: ResourceWarning: unclosed file <_io.BufferedReader name='/home/dav/.local/share/cartopy/shapefiles/natural_earth/physical/10m_coastline.shx'>
  geometries = tuple(shapereader.Reader(path).geometries())
/home/dav/anaconda3/envs/cartopy/lib/python3.6/site-packages/cartopy/feature.py:184: ResourceWarning: unclosed file <_io.BufferedReader name='/home/dav/.local/share/cartopy/shapefiles/natural_earth/physical/10m_coastline.dbf'>
  geometries = tuple(shapereader.Reader(path).geometries())
ibcao: closing map.
ok
test_map_depth_vs_interp_depth (test_depth_map_depth_interp.IbcaoDepthTest) ... ibcao read, shape: (11617, 11617)
INFO:root:testing map_depth vs interp_depth
setting up interpolation function..
ibcao: closing map.
/home/dav/anaconda3/envs/cartopy/lib/python3.6/site-packages/scipy/io/netcdf.py:299: RuntimeWarning: Cannot close a netcdf_file opened with mmap=True, when netcdf_variables or arrays referring to its data still exist. All data arrays obtained from such files refer directly to data on disk, and must be copied before the file can be cleanly closed. (See netcdf_file docstring for more information on mmap.)
  ), category=RuntimeWarning)
ok
test_make_test_map (test_plot.IbcaoPlotTest) ... ibcao read, shape: (11617, 11617)
INFO:root:make test map
ibcao: closing map.
ok
test_template (test_plot.IbcaoPlotTest) ... ibcao read, shape: (11617, 11617)
INFO:root:testing template
ibcao: closing map.
ok
test_corners (test_proj.IbcaoProjTest) ... ibcao read, shape: (11617, 11617)
INFO:root:test corners
ibcao: closing map.
/home/dav/anaconda3/envs/cartopy/lib/python3.6/site-packages/scipy/io/netcdf.py:299: RuntimeWarning: Cannot close a netcdf_file opened with mmap=True, when netcdf_variables or arrays referring to its data still exist. All data arrays obtained from such files refer directly to data on disk, and must be copied before the file can be cleanly closed. (See netcdf_file docstring for more information on mmap.)
  ), category=RuntimeWarning)
ok
test_ibcao_grid (test_proj.IbcaoProjTest) ... ibcao read, shape: (11617, 11617)
INFO:root:test grid coordinates
ibcao: closing map.
/home/dav/anaconda3/envs/cartopy/lib/python3.6/site-packages/scipy/io/netcdf.py:299: RuntimeWarning: Cannot close a netcdf_file opened with mmap=True, when netcdf_variables or arrays referring to its data still exist. All data arrays obtained from such files refer directly to data on disk, and must be copied before the file can be cleanly closed. (See netcdf_file docstring for more information on mmap.)
  ), category=RuntimeWarning)
ok
test_north_pole (test_proj.IbcaoProjTest) ... ibcao read, shape: (11617, 11617)
INFO:root:test north pole
ibcao: closing map.
/home/dav/anaconda3/envs/cartopy/lib/python3.6/site-packages/scipy/io/netcdf.py:299: RuntimeWarning: Cannot close a netcdf_file opened with mmap=True, when netcdf_variables or arrays referring to its data still exist. All data arrays obtained from such files refer directly to data on disk, and must be copied before the file can be cleanly closed. (See netcdf_file docstring for more information on mmap.)
  ), category=RuntimeWarning)
ok
test_np_stere (test_proj.IbcaoProjTest) ... ibcao read, shape: (11617, 11617)
INFO:root:testing np stereographic vs our projection
ibcao: closing map.
/home/dav/anaconda3/envs/cartopy/lib/python3.6/site-packages/scipy/io/netcdf.py:299: RuntimeWarning: Cannot close a netcdf_file opened with mmap=True, when netcdf_variables or arrays referring to its data still exist. All data arrays obtained from such files refer directly to data on disk, and must be copied before the file can be cleanly closed. (See netcdf_file docstring for more information on mmap.)
  ), category=RuntimeWarning)
ok
```
