### 2017-05-03 #250 @dvalters
Hi @gauteh, my review is below:

## Software functionality and installation

I was able to download and install the software, although really it is just a script rather than a package that can be installed using a setup script or package manager. (Though I notice you are working on one in the master branch? If you could include this in the release it would improve this software by automating the installation.) The dependencies aren't listed explicitly (such as in a `requirements.txt` file). It's implied of course that you need to install Cartopy, Matplotlib, and the associated dependencies with those packages. Stating the dependencies explicitly would improve the software package.

The test command in the README.md produced the figure as indicated:
![figure_1-5](https://cloud.githubusercontent.com/assets/10996571/25684000/c73f62d6-3056-11e7-84ed-c1ceeceed666.png).

It might be nice to see this as an example in your accompanying paper to highlight the quality of figures this software can produce.

The unit tests provided in the `/tests` folder ran successfully, though with warnings about unclosed files which may want to be addressed. (I raised an issue here: gauteh/ibcao_py#1, though it's not essential to fix this for acceptance.) The unit testing shell script is not documented anywhere though. It was a positive to see unit testing being employed. I was a bit puzzled by what some of the test output plots where meant to represent, and couldn't tell if they were correct or not. For example:

![depth_rectbivariate](https://cloud.githubusercontent.com/assets/10996571/25683396/e2524e24-3053-11e7-8f00-4ae18fcfe058.png)

## Documentation
The documentation is sparse. The only documentation I could find was the README.md file, which was really just a brief set up guide and an example, rather than detailed documentation of the API. I would say this is the main issue holding back the software package from being accepted.

There are no community guidelines provided (see the checkbox for details of what is suggested)

## Paper
The paper is very brief (only a single short paragraph). I understand the JOSS papers are only meant be short introductions to the software, but even so, it could be expanded on a little bit further. Who are the target audience? Oceanographers, presumably, but are there others? Earth Scientists in general?

You mention the test script 'showcases some of the possibilities', but you could briefly list the main ones here to improve the paper for the target audience. Does it do anything other than project data to one specific projection system?

### References

References are encouraged to have a DOI where available. Matplotlib is missing a DOI in your `.bib` file, but there is one provided here that you could include: https://matplotlib.org/citing.html. I couldn't find a DOI for Cartopy, but the format of the reference should list the author field as `Met Office`, as per this bibtex example: http://scitools.org.uk/cartopy/docs/v0.15/citation.html

## Licence file
Renaming it LICENCE (or LICENSE, depending on your spelling preference) instead of COPYING may make it more clear that this is the licence file.

## Summary
The software provides useful and novel functionality to a scientific community. I think though it needs major revisions, mainly because of the lack of proper documentation for the API. (The user shouldn't have to read through all the unit tests to figure out what the software can do - sell it to them through good documentation and the paper!)  I think once the documentation and the paper have been revised, it can be accepted, and I will tick off the remaining checkboxes.
