### 2017-05-04 gauteh/ibcao_py#1 @gauteh
Hi Declan,

the warning stems from the map-file being memorymapped when plotting or
using the map to e.g. interpolate depth. Variables in the class are
referencing the map-data and the file can therefore not be closed,
secondly, there is no way for the Ibcao-class to know whether any
variables are referencing the data or even know when it is safe to close
the file.

We either have to copy the full map to memory, which is significant and
probably time-consuming in this case, or we have to live with the
warning which essentialy is Scipy preventing the file from being closed
as long as there are arrays pointing to the file. As far as I can
understand, the file will still be left open if there are references to
the mmaped file.

Relevant issues:
* https://github.com/scipy/scipy/issues/3630#issuecomment-53982155
* https://docs.scipy.org/doc/scipy-0.16.1/reference/generated/scipy.io.netcdf.netcdf_file.html
