### 2017-05-05 gauteh/ibcao_py#1 @dvalters
Hi Gauteh, thanks for explaining this - I understand a lot better now why the warnings arise and that it's related to scipy.

Best wishes,
Declan
