### 2017-05-16 fec4439
README: contributions guidelines
