+++ README.md
@@ -34,44 +34,7 @@ $ python setup.py install
 $ python ibcao.py
 ```
 
-3. or in your code do something like:
-```python
-  print ("testing ibcao class")
-  import matplotlib.pyplot as plt
-  import matplotlib.cm as cm
-  import cartopy.crs as ccrs
-  from ibcao import *
-  from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
-
-  i = IBCAO ()
-
-  f = i.template ()
-
-  # lets put some text along the parallels
-  lat = np.arange (65, 90, 5)
-  lon = np.repeat (0, len(lat))
-
-  # regular lat, lon projection
-  g = ccrs.Geodetic ()
-
-  for lon, lat in zip (lon, lat):
-    plt.text (lon, lat, LATITUDE_FORMATTER.format_data(lat), transform = g)
-
-  # and some along the meridians
-  lon = [-45, 45, 135, -135]
-  lat = np.repeat (70, len(lon))
-
-  for lon, lat in zip (lon, lat):
-    plt.text (lon, lat, LONGITUDE_FORMATTER.format_data(lon), transform = g)
-
-  # also; the north pole
-  plt.text (0, 90, "NP", transform = g)
-
-  plt.show ()
-```
-
-> check out the test cases in `tests/` for some inspiration on how to use the
-> class.
+3. Have a look at the documentation for how to [get started with the package](https://github.com/gauteh/ibcao_py/blob/master/doc/IBCAO%20demonstration.ipynb).
 
 ## Contributing or reporting issues
 
