### 2017-05-16 656cb63
readme: refer to docs
