### 2017-05-16 def5da0
readme: fix typos
