+++ ibcao/ibcao.py
@@ -38,14 +38,14 @@ class IBCAO:
 
   # look in current path for grid file, I usually symlink it in here
   # from somewhere.
-  ibcao_grid_name = 'IBCAO_V3_500m_RR.grd'
-  ibcao_grid = os.path.join(os.path.dirname(os.path.realpath(__file__)), ibcao_grid_name)
+  _ibcao_grid_name = 'IBCAO_V3_500m_RR.grd'
+  _ibcao_grid = os.path.join(os.path.dirname(os.path.realpath(__file__)), _ibcao_grid_name)
 
   VERSION = '3.0'
   REFERENCE = 'Jakobsson, M., L. A. Mayer, B. Coakley, J. A. Dowdeswell, S. Forbes, B. Fridman, H. Hodnesdal, R. Noormets, R. Pedersen, M. Rebesco, H.-W. Schenke, Y. Zarayskaya A, D. Accettella, A. Armstrong, R. M. Anderson, P. Bienhoff, A. Camerlenghi, I. Church, M. Edwards, J. V. Gardner, J. K. Hall, B. Hell, O. B. Hestvik, Y. Kristoffersen, C. Marcussen, R. Mohammad, D. Mosher, S. V. Nghiem, M. T. Pedrosa, P. G. Travaglini, and P. Weatherall, The International Bathymetric Chart of the Arctic Ocean (IBCAO) Version 3.0, Geophysical Research Letters, doi: 10.1029/2012GL052219.'
 
   # previously kept in ibcao.cpt
-  COLORMAP = """\
+  _COLORMAP = """\
 # downloaded from IBCAO homepage
 #Discrete color table for Ocean and continous for land in RGB for the Arctic bathymetry and topography
 -6000	18	10	59	-5000	18	10	59
@@ -78,7 +78,7 @@ class IBCAO:
 1500	170	170	170	5000	200	200	200
 """
 
-  def __init__ (self, ibcao_grd_file = ibcao_grid):
+  def __init__ (self, ibcao_grd_file = _ibcao_grid):
     self.ibcao_grid = ibcao_grd_file
     if not os.path.exists (self.ibcao_grid):
       print ('IBCAO grid could not be found in:' + self.ibcao_grid + ' , download from: http://www.ngdc.noaa.gov/mgg/bathymetry/arctic/grids/version3_0/IBCAO_V3_500m_RR.grd')
@@ -116,17 +116,23 @@ class IBCAO:
 
     # store for short-cut
     self.g = ccrs.Geodetic ()
+    self._ups = self.get_cartopy ()
 
     # don't close when mmapped: scipy#3630
     #self.ibcao_nc.close ()
 
   def close (self):
+    """
+    Closes the map file. The map is memorymapped, so this will cause a warning unless all references to the map have been removed.
+    """
     # make sure you don't close in case mmap is used elsewhere
     print ("ibcao: closing map.")
     self.ibcao_nc.close ()
 
   def get_cartopy (self):
     """
+    Deprecated: Use `ups`.
+
     Returns a Cartopy instance set up for the IBCAO UPS variant. A separate
     instance is used internally.
     """
@@ -138,7 +144,17 @@ class IBCAO:
 
     return m
 
-  def get_proj_str (self):
+  @property
+  def ups (self):
+    """
+    Returns a Cartopy Stereographic instance set up for the Universal Polar
+    Stereographic (UPS) projection used in the IBCAO.
+    """
+    return self._ups
+
+
+  @property
+  def proj_str (self):
     """
     Returns a Proj.4 string for the IBCAO UPS variant.
     """
@@ -165,13 +181,18 @@ class IBCAO:
         'y0' : 0
         }
 
-  def get_proj (self):
+  @property
[…]
