+++ doc/ibcao.rst
@@ -1,0 +1,22 @@
+ibcao package
+=============
+
+Submodules
+----------
+
+ibcao\.ibcao module
+-------------------
+
+.. automodule:: ibcao.ibcao
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: ibcao
+    :members:
+    :undoc-members:
+    :show-inheritance:
