+++ doc/index.rst
@@ -1,0 +1,23 @@
+.. IBCAO_py documentation master file, created by
+   sphinx-quickstart on Tue May 16 12:11:12 2017.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to IBCAO_py's documentation!
+====================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   modules
+
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
