+++ doc/modules.rst
@@ -1,0 +1,7 @@
+ibcao
+=====
+
+.. toctree::
+   :maxdepth: 4
+
+   ibcao
