### 2017-05-16 f0fbcf1
doc: set up for readthedocs
reference included

empty requir for docs

add module dir

cartopy 15

cartopy

setuptools

reorder req

readthedocs: use conda

add scitools channel
