+++ .readthedocs.yml
@@ -1,0 +1,5 @@
+conda:
+  file: environment.yml
+python:
+  version: 3
+
