+++ paper/ibcao_py.bib
@@ -1,17 +1,25 @@
-@online{Cartopy,
-  author = {SciTools},
-  title = {Cartopy},
-  year = 2017,
-  url = {http://scitools.org.uk/cartopy/},
-  urldate = {2017-04-07}
+@manual{Cartopy,
+author = {{Met Office}},
+title = {Cartopy: a cartographic python library with a matplotlib interface},
+year = {2010 - 2015},
+address = {Exeter, Devon },
+url = {http://scitools.org.uk/cartopy}
 }
 
-@online{matplotlib,
-  author = {The Matplotlib development team},
-  title = {Matplotlib},
-  year = 2017,
-  url = {http://matplotlib.org},
-  urldate = {2017-04-07}
+@Article{matplotlib,
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing In Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python
+  for application development, interactive scripting, and
+  publication-quality image generation across user
+  interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi = {10.1109/MCSE.2007.55},
+  year      = 2007
 }
 
 @article{Jakobsson2012,
