+++ setup.py
@@ -23,7 +23,7 @@ setup(
     # Versions should comply with PEP440.  For a discussion on single-sourcing
     # the version across setup.py and the project code, see
     # https://packaging.python.org/en/latest/single_source_version.html
-    version='1.0',
+    version='1.1',
 
     description='A matplotlib library for using the International Bathymetric Chart of the Arctic Ocean with cartopy and matplotlib',
     long_description=long_description,
