+++ doc/conf.py
@@ -58,7 +58,7 @@ author = 'Gaute Hope'
 # built documents.
 #
 # The short X.Y version.
-version = '1.0'
+version = '1.1'
 # The full version, including alpha/beta/rc tags.
 release = version 
 
