### 2017-05-17 #250 @gauteh
Arfon Smith writes on mai 17, 2017 17:22:
> @gauteh - At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
>

Hi Arfon,

the DOI is: 10.5281/zenodo.580775  (https://zenodo.org/record/580775)

- gaute
