### 2017-05-22 ebf361a
JOSS paper updates
