### 2017-05-22 #250 @gauteh
@arfon: great! many thanks!

@dvalters: Thanks for the thorough review!

Arfon Smith writes on mai 22, 2017 18:52:
> @dvalters - many thanks for your review here ✨
>
> @gauteh - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00250 ⚡️ 🚀 💥
